requirements/requirements-fastapi.txt
mlflow==3.11.1
numpy==2.4.4
scipy==1.17.1
pybaselines==1.2.1
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
requests==2.32.3
python-multipart==0.0.26 
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
prometheus-client==0.25.0
prometheus-fastapi-instrumentator==7.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
pandas==2.2.3
